redefine the meaning of report.getRootException()
Before, this returned the head of the full exception chain, whose tail was the root cause.
Now, this just returns the root cause.

The main motivation is that Errbit really does not handle well exception chains.  The web page UI handles it ok, but the email updates it sends attempts to condense the backtrace, and the chaining makes it non-sensical.

The full exception chain will be recorded in the ‘ALL THROWN EXCEPTIONS’ environment/cgi variable.