add jax-ws soap handler
This will make sure soap faults trigger an error report, and that it will be correctly indicated a 'user error' if the soap fault code warrants it.

This also adds an arquillian test that uses some of the same infrastructure as the jax-rs test.
The test also checks the usage of the @Action interceptor, which was added mainly for jax-ws support.