make unconfigured parameter sanitizer behavior consistent
Previously, with CDI, failure to configure a ParameterSanitizer would fail the deployment. Without CDI, it would deploy but every parameter would be removed.
Both behaviors are acceptable, but it'd be better if they were consistent.
Since the first behavior can't be emulated without CDI, we'll go with the second behavior.
Plus, it is arguably more user-friendly.

This is implemented by making @Fallback a project-level qualifier, not just for the recording subpackages.