add ability for an app to use ServletRequest.getInputStream()
Currently, red egg will prevent this, due to its calling ServletRequest.getParameterMap(); for x-www-form-urlencoded POSTs, this method chews up the input stream and makes it unavailable to the app.

This commit adds the ability to instruct red egg to stop doing this for a configurable set of url paths. It is also possible to use other types of selection criteria, by implementing a custom RequestMatcher (perhaps by selecting only POST requests, and not GET/DELETE etc. requests).