Make environment variables and system properties NOT passed by default
It's not generally known what data are stored in properties and environment variables, so it should be up to the developer on a given project whether or not to include them.  The new default behavior is that developers must opt in to including these in error reports.

It should be noted that calling the explicit `recordSystemProperties` and `recordEnvironmentVariables` methods on `ErrorRecorder` is an opt-in and sets the respective boolean to true for correctness.