avoid race condition in EndToEnd tests
We had been mitigating this somewhat by a 500 ms wait; but this isn't always enough (and on the CI server, it seems like it is never enough).
Now, the test just keeps hitting the API until it has received the report.

Also, this adds some synchronization to the dummy errbit api that should have been there from the beginning.