handle RejectedExecutionExceptions in error report queue
This gracefully handles cases where InMemoryErrorQueue cannot add a new report to the queue because the ExecutorService can't handle it.
Before, the RejectedExecutionException would propagate, and hide the original exception.

This happened a few weeks ago in the IDM apps,
when they began getting `java.lang.OutOfMemoryError: unable to create new native thread` failures.
(There were too many running threads on the OS, I believe.)
This caused `ExecutorService.submit()` to fail, and the original problem was not logged
(only the RejectedExecutionException).