load RedEggVersion eagerly
Before, the version lookup would happen when the first notification is sent.
However, in the give-ep app, at this point it seems the thread context classloader was set such that the XPath implementation loaded was not the JDK's, but an apache one, and it didn't have access to the right classes it needed, and this resulted in a linkage error.

So rather than being more careful about which XPath implementation is loaded, which I imagine is tricky, I worked around the problem my moving the XPath query early in the app sequence, when the TCCL is 'pure'.