improve request timing info
It was annoying to have the ‘finish’ attribute so far from the ‘start’ attribute in the rollbar UI.
(And rollbar sorts these alphabetically, so ‘finished’ was listed first.)

Now, these are printed next to each other.
Also the duration between the two (which I used to often calculate mentally) is now printed as well.
The ISO 8601 formatting for a duration is a little funny (e.g. ‘PT11.509S’, which means 11.509 seconds), but it’s good enough.