record/report the query string
This is useful for 'replaying' the request from rollbar's UI
(either using the button, or copy/pasting the curl request.)

This is a little tricky because the query string should be sanitized/scrubbed
if it contains sensitive information.
This commit uses the existing framework for sanitizing query parameters,
but since there's no guarantee that the sanitizer will return a list of the same size,
some extra complexity is required to handle this gracefully.