record MDC attributes on warning logs, as well
This should have been done when I made red-egg support rollbar warnings, but I missed it.

I'm actually not sure that I shouldn't be recording MDC attributes for any log level; I don't really know if that could cause too much noise or if it'd be useful. So I'll leave a limit in for now.