Implement monitoring for stuck requests
Several times I've been burned by requests getting permanently stuck
on a network call (when no socket read timeout was configured),
which chews up a thread and potentially other resources.
Most recently this happened in give-ep-cortex
when OBIEE had some major issues,
but I have had it happen in other contexts before.

Monitoring for these in a useful way is somewhat complex,
but I think it's worth it.

This commit doesn't make the timeout config configurable except in CDI mode,
which was needed for the tests.
I don't think we'll need to configure the timeouts, in practice;
the default will probably be sufficient.

The default timeout is pretty conservative (30 m)
because I don't want false-positive notifications to become an issue.
(I have made that mistake before.)
If requests take a long time, but are not truly stuck,
they will show up in APM/metrics and such.
This monitoring is intended for truly hung requests,
like bad blocking IO and infinite loops,
not just abnormally slow ones.
(Hung requests will not show up in APM.)

The request details captured are not as rich,
but should be sufficient given that stuck threads are relatively rare.

This commit also adds a shutdownNow() call to executor services
that are closed at app shutdown.
May as well try to get the executor to shut down faster.