fix marshalling error in JaxwsRecordingIntegrationTest
It turns out that the 'throws SoapFaultException' declaration was triggering the error.
It's optional, since it's a runtime exception, but I had put it there for documentation purposes.

The problem seems to be that CXF treats it like a normal fault exception, which usually is required to have a no-arg constructor. SoapFaultException doesn't have one.

Removing the declaration makes the CXF client work as I originally expected it to.