beef up this test
This changes the project to use java 8 for the tests, but still use java 7 for compiling the red-egg jar.
See https://gist.github.com/aslakknutsen/9648594
However, I didn't want to add the complexity of the animal sniffer plugin and such.
I suspect I'll be fully upgrading this project to java 8 in the near future; java 8 is getting old at this point.