remove untested support for semicolon query-string separators
There's an old W3C recommendation [1] that web services support semicolons as an
alternative to ampersands.

However, tomcat doesn't support it [2], it looks to me like undertow doesn't
either [3], and it's certainly not in the servlet spec [4].
I've also read somewhere that the most up-to-date W3C recommendation
(covering html5) doesn't include the the semicolon text.

So, I believe this is more likely to cause harm than good, and I'm removing it.

[1]: https://www.w3.org/TR/1999/REC-html401-19991224/appendix/notes.html#h-B.2.2
[2]: https://bz.apache.org/bugzilla/show_bug.cgi?id=39740
[3]: https://github.com/undertow-io/undertow/blob/2.0.0.Final/core/src/main/java/io/undertow/util/QueryParameterUtils.java
[4]: http://download.oracle.com/otn-pub/jcp/servlet-4-final-eval-spec/servlet-4_0_FINAL.pdf