add note about incorrectness in some cases
If a request has both a 'foo=bar' query string parameter and a 'foo=baz' post body
parameter, the servlet container will put both into the parameter map.
The code will currently classify both of these as query string parameters, which is wrong.
But it's not likely to bother anyone so I'm leaving it for now.