Add support for capturing error-details links
These links will be logged and sent in http response headers.

We use an extension relation type since there isn't a suitable registered one
(I looked, and https://www.iana.org/assignments/link-relations/link-relations.xhtml
doesn't have anything useful.)

Note: the addHeader() call RedEggFilter doesn't take affect
if an exception is thrown past the filter.
I didn't research it, but I am guessing that the servlet container
needs to throw away the headers along with the body
when it aborts a request and renders the (possibly default) error page.
Most of Cru's apps don't let exceptions propagate out to the container,
so I am not worrying about this caveat.
The IT tests in this project, though, mostly do propagate exceptions,
and so testing for the Link header must be done in the logging test
(which doesn't propagate exceptions.)

This commit adds a little bit of refactoring to the arquillian dependency lists
for the integration tests, since LoggingReporter now brings in another class
(it transitively needs ErrorLink).