Add log4j 2.x support
I want log4j 2 to be preferred, because there's a decent chance that an
app might have both 1.x and 2.x on the classpath if it is bridging from 1.x to 2.x.

But this means the end-to-end test can't have log4j 2.x on the classpath,
or else the test needs to use log4j 2.x.
I decided to go with the former.