Use MediaType API go get charset
(instead of parsing the Content-Type header.)

This is not only easier to read, it is also less brittle.
I don't remember details exactly, but a few months ago I saw a failure here,
because there was no space before the 'charset' parameter.
I don't recall if the HTTP spec allows this or not,
but it's better to leave that sort of thing to jax-rs regardless.