Bump a bunch of dependencies suggested by Intellij
But not the Wildfly-related ones, since those should probably be done together.

Mockito and Logback can't be updated to the most recent releases
because they drop java 8 support.

Objenesis, used by Mockito, needed to be downgraded from the default to 3.1
because version 3.3 causes problems on older versions of Wildfly.
Here's a rough description of the problem, but it's slightly off-topic
because it's about Spring and not Objenesis:
https://stackoverflow.com/questions/65100927/springboot-wildfly-10-deployment-error-jdk-unsupported-module-not-found/66176378#66176378
The answer is helpful, though, because Spring only has the problem because it internally uses Objenesis.