Remove Joda Time
This is way overdue; java.time/jsr-310 has been here for over a decade,
and we dropped java 7 support in this project in I think 2017.

One CDI consequence is that the Clock now shouldn't use @Default, because
Clock is no longer a red-egg type but a JDK type, and clients might
want to define their own producer for @Default Clock.
If they do, we'll use it, but if not, then we'll use the system clock.

This also revealed an Arquillian-related problem in the EndToEndWithCdi test;
the packaging was including some red-egg classes in the webapp in addition
to the red-egg jar, and so this commit fixes that.

This commit also fixes some unrelated unused import problems.