Reverse the order of exceptions in trace chains
The default ordering of rollbar's RollbarThrowableWrapper puts outer exceptions first
and inner exceptions later.
This mimics the behavior of Throwable.printStackTrace().
As a result, the Rollbar UI puts outer exceptions at the top of the page
and inner exceptions further down the page
(also like Throwable.printStackTrace()).

But generally, the most inner exception (the root cause)
is the one we really care about.
So this commit reverses the list of throwables,
which causes the Rollbar UI to put the most inner exception at the top of the page.
As near as I can tell, the Rollbar UI does not do anything else interesting
that takes into consideration the order of the throwables.
As a side note, this is not true with Throwable.printStackTrace();
that method will ensure that stack frames that are identical between
an inner exception and an outer exception are only printed once,
and get a '... 3 more' indicator for the duplicated frames.
But the Rollbar UI does not help by collapsing identical stack frames this way.